+++ src/Coupons/Coupon.php
@@ -56,7 +56,7 @@ class Coupon implements Contract
         }
 
         if ($this->has('redeemed') && $this->has('maximum_uses')) {
-            if ($this->has('redeemed') >= $this->get('maximum_uses')) {
+            if ($this->get('redeemed') >= $this->get('maximum_uses')) {
                 return false;
             }
         }