+++ src/Exceptions/GatewayCallbackMethodDoesNotExist.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace DoubleThreeDigital\SimpleCommerce\Exceptions;
+
+class GatewayCallbackMethodDoesNotExist extends \Exception
+{
+    //
+}