+++ src/Http/Controllers/CouponController.php
@@ -28,6 +28,7 @@ class CouponController extends BaseActionController
 
     public function destroy(DestroyRequest $request)
     {
+        // TODO: We need to figure out a way of making this work with different drivers (eg. Eloquent uses coupon_id instead of coupon)
         $this->getCart()
             ->data([
                 'coupon' => null,