+++ src/Orders/Order.php
@@ -93,7 +93,7 @@ class Order implements Contract
             return $this;
         }
 
-        if (! $this->has('customer')) {
+        if (! $this->has('customer') || $this->get('customer') === null) {
             return null;
         }
 
@@ -106,6 +106,10 @@ class Order implements Contract
             $this->set('coupon', $coupon);
 
             return $this;
+        }
+
+        if (! $this->has('coupon') || $this->get('coupon') === null) {
+            return null;
         }
 
         return Coupon::find($this->get('coupon'));