+++ src/ServiceProvider.php
@@ -24,6 +24,7 @@ class ServiceProvider extends AddonServiceProvider
     ];
 
     protected $fieldtypes = [
+        Fieldtypes\CountryFieldtype::class,
         Fieldtypes\MoneyFieldtype::class,
         Fieldtypes\ProductVariantFieldtype::class,
         Fieldtypes\ProductVariantsFieldtype::class,
@@ -34,8 +35,18 @@ class ServiceProvider extends AddonServiceProvider
             Listeners\EnforceBlueprintFields::class,
         ],
         Events\OrderPaid::class => [
-            Listeners\SendOrderPaidNotifications::class,
+            Listeners\SendConfiguredNotifications::class,
         ],
+        Events\StockRunningLow::class => [
+            Listeners\SendConfiguredNotifications::class,
+        ],
+        Events\StockRunOut::class => [
+            Listeners\SendConfiguredNotifications::class,
+        ],
+    ];
+
+    protected $modifiers = [
+        Modifiers\Currency::class,
     ];
 
     protected $routes = [