+++ src/Tags/CouponTags.php
@@ -16,26 +16,22 @@ class CouponTags extends SubTag
             return [];
         }
 
-        $coupon = isset($this->getCart()->data['coupon'])
-            ? $this->getCart()->data['coupon']
-            : null;
+        $coupon = $this->getCart()->coupon();
 
-        if ($coupon === null) {
+        if (! $coupon) {
             return [];
         }
-
-        $coupon = Coupon::find($coupon);
 
         return $coupon->toAugmentedArray();
     }
 
     public function has()
     {
-        if (!$this->hasCart()) {
+        if (! $this->hasCart()) {
             return false;
         }
 
-        return isset($this->getCart()->data['coupon']);
+        return ! is_null($this->getCart()->coupon());
     }
 
     public function redeem()